+++ plugins/Contents/tests/System/ContentsTest.php
@@ -7,9 +7,9 @@
  */
 namespace Piwik\Plugins\Contents\tests\System;
 
+use Piwik\Tests\Framework\Fixture;
 use Piwik\Tests\Framework\TestCase\SystemTestCase;
 use Piwik\Plugins\Contents\tests\Fixtures\TwoVisitsWithContents;
-use Piwik\Translate;
 
 /**
  * Testing Contents
@@ -48,13 +48,13 @@ class ContentsTest extends SystemTestCase
     protected function setup()
     {
         parent::setup();
-        Translate::loadAllTranslations();
+        Fixture::loadAllTranslations();
     }
 
     protected function tearDown()
     {
         parent::tearDown();
-        Translate::reset();
+        Fixture::resetTranslations();
     }
 
     public function getApiForTesting()