+++ plugins/LanguagesManager/LanguagesManager.php
@@ -15,11 +15,9 @@ use Piwik\Common;
 use Piwik\Config;
 use Piwik\Container\StaticContainer;
 use Piwik\Cookie;
-use Piwik\Db;
 use Piwik\Intl\Locale;
 use Piwik\Piwik;
 use Piwik\ProxyHttp;
-use Piwik\Translate;
 use Piwik\Translation\Translator;
 use Piwik\View;
 
@@ -156,7 +154,7 @@ class LanguagesManager extends \Piwik\Plugin
             $languageCode = Common::extractLanguageCodeFromBrowserLanguage(Common::getBrowserLanguage(), API::getInstance()->getAvailableLanguages());
         }
         if (!API::getInstance()->isLanguageAvailable($languageCode)) {
-            $languageCode = Translate::getLanguageDefault();
+            $languageCode = StaticContainer::get('Piwik\Translation\Translator')->getDefaultLanguage();
         }
         return $languageCode;
     }