+++ plugins/LanguagesManager/TranslationWriter/Filter/EncodedEntities.php
@@ -9,7 +9,7 @@
 
 namespace Piwik\Plugins\LanguagesManager\TranslationWriter\Filter;
 
-use Piwik\Translate;
+use Piwik\Container\StaticContainer;
 
 class EncodedEntities extends FilterAbstract
 {
@@ -38,12 +38,12 @@ class EncodedEntities extends FilterAbstract
             foreach ($pluginTranslations as $key => $translation) {
 
                 if (isset($this->baseTranslations[$pluginName][$key]) &&
-                    $this->baseTranslations[$pluginName][$key] != Translate::clean($this->baseTranslations[$pluginName][$key])) {
+                    $this->baseTranslations[$pluginName][$key] != StaticContainer::get('Piwik\Translation\Translator')->clean($this->baseTranslations[$pluginName][$key])) {
                     continue; // skip if base translation already contains encoded entities
                 }
 
                 // remove encoded entities
-                $decoded = Translate::clean($translation);
+                $decoded = StaticContainer::get('Piwik\Translation\Translator')->clean($translation);
                 if ($translation != $decoded) {
                     $this->filteredData[$pluginName][$key] = $translation;
                     $translations[$pluginName][$key] = $decoded;