+++ tests/PHPUnit/Integration/Plugin/ArchivedMetricTest.php
@@ -12,13 +12,10 @@ use Piwik\Columns\Dimension;
 use Piwik\DataTable;
 use Piwik\Metrics\Formatter;
 use Piwik\Plugin\ArchivedMetric;
-use Piwik\Plugin\ComputedMetric;
-use Piwik\Plugins\UserCountry\Columns\City;
 use Piwik\Plugins\UserCountry\Columns\Country;
 use Piwik\Site;
 use Piwik\Tests\Framework\Fixture;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
-use Piwik\Translate;
 
 /**
  * @group ArchivedMetric
@@ -40,7 +37,7 @@ class ArchivedMetricTest extends IntegrationTestCase
     {
         parent::setUp();
 
-        Translate::loadEnglishTranslation();
+        Fixture::loadAllTranslations();
 
         Fixture::createWebsite('2015-01-01 00:00:00');
 
@@ -50,7 +47,7 @@ class ArchivedMetricTest extends IntegrationTestCase
 
     public function tearDown()
     {
-        Translate::unloadEnglishTranslation();
+        Fixture::resetTranslations();
         parent::tearDown();
     }